requirements.txt
gunicorn==20.1.0
flask==2.3.2
PyJWT==2.7.0
mysql-connector-python==8.0.33
mysql==0.0.3
mysqlclient==2.1.1
stripe==